+++ src/components/FormRender.ts
@@ -1,5 +1,0 @@
-import { h } from 'vue'
-
-export default function Render(props: any) {
-  return props.render(props.formItem, h)
-}