+++ vite.config.ts
@@ -3,6 +3,7 @@ import viteSvgIcons from 'vite-plugin-svg-icons'
 import path from 'path'
 import { defineConfig } from 'vite'
 import dotenv from 'dotenv'
+import vueJsx from '@vitejs/plugin-vue-jsx'
 // 在正式打包的时候，可以把这两行代码放开
 // import Components from 'unplugin-vue-components/vite'
 // import { ArcoResolver } from 'unplugin-vue-components/resolvers'
@@ -30,6 +31,7 @@ export default defineConfig(({ mode }) => {
         iconDirs: [path.resolve(process.cwd(), 'src/icons')],
         symbolId: 'icon-[dir]-[name]',
       }),
+      vueJsx(),
       // 在正式打包的时候，可以把这三行代码放开
       // Components({
       //   resolvers: [ArcoResolver()],